# expected-oas/contextapi.openapi.json
{
  "components": {
    "schemas": {
...
      "ImageMetadata": {
        "description": "The ImageMetadata that can be requested for each uploaded image",
        "properties": {
          "created": {
            "$ref": "#/components/schemas/SystemTime",
            "description": "The time when the image was first stored"
          },
          "file_name": {
            "description": "The user specified file name of the image",
            "type": "string"
          },
          "size": {
            "description": "The size of the image in bytes",
            "format": "uint64",
            "minimum": 0.0,
            "type": "integer"
          }
        },
        "required": [
          "created",
          "file_name",
          "size"
        ],
        "type": "object"
      },
...
      "Sha256Hash": {
        "description": "Wraps a String and checks whether it is a valid SHA256 hash on construction and deserialization.",
        "type": "string"
      },
...
      "SystemTime": {
        "properties": {
          "nanos_since_epoch": {
            "format": "uint32",
            "minimum": 0.0,
            "type": "integer"
          },
          "secs_since_epoch": {
            "format": "uint64",
            "minimum": 0.0,
            "type": "integer"
          }
        },
        "required": [
          "nanos_since_epoch",
          "secs_since_epoch"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The REST-API to access the Hardware as a Service implementation.",
    "title": "HWaaS Context API",
    "version": "5.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/images": {
      "get": {
        "description": "Returns a list of all available images currently stored",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "format": "uint64",
                    "minimum": 0.0,
                    "type": "integer"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Return a dictionary containing the image hash and image size in byte"
          },
          "500": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "Internal error reason"
          }
        },
        "summary": "list images",
        "tags": [
          "Image API"
        ]
      },
      "post": {
        "description": "Upload an image",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "array"
              }
            }
          },
          "description": "multipart form data",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "plain text"
          },
          "500": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "Internal error reason"
          }
        },
        "summary": "Upload image",
        "tags": [
          "Image API"
        ]
      }
    },
    "/images/{image_hash}": {
      "delete": {
        "description": "Mark an image for garbage collection",
        "parameters": [
          {
            "description": "hash of a boot image (sha256sum)",
            "in": "path",
            "name": "image_hash",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Sha256Hash",
              "description": "hash of a boot image (sha256sum)"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "202": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "Image marked for garbage collection"
          }
        },
        "summary": "delete",
        "tags": [
          "Image API"
        ]
      },
      "get": {
        "description": "Request metadata of an image",
        "parameters": [
          {
            "description": "hash of a boot image (sha256sum)",
            "in": "path",
            "name": "image_hash",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Sha256Hash",
              "description": "hash of a boot image (sha256sum)"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageMetadata"
                }
              }
            },
            "description": "metadata"
          },
          "500": {
            "content": {
              "text/plain; charset=utf-8": {}
            },
            "description": "Internal error reason"
          }
        },
        "summary": "Get image metadata",
        "tags": [
          "Image API"
        ]
      }
    },
...
  }
}